requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.5.1
minio==7.2.7
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.4
pydantic-settings==2.3.3
